core code updates - minor enhancements
git-svn-id: https://svn.code.sf.net/p/hsqldb/svn/base/trunk@5552 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ src/org/hsqldb/persist/TextCache.java
@@ -450,7 +450,7 @@ public class TextCache extends DataFileCache {
         }
     }
 
-    public int getLineNumber() {
+    public long getLineNumber() {
         return ((RowInputText) rowIn).getLineNumber();
     }
