+++ src/org/hsqldb/persist/TextFileReader.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2001-2011, The HSQL Development Group
+/* Copyright (c) 2001-2016, The HSQL Development Group
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -310,7 +310,7 @@ public class TextFileReader {
         return header;
     }
 
-    public int getLineNumber() {
+    public long getLineNumber() {
         return ((RowInputText) rowIn).getLineNumber();
     }
 }
