+++ src/org/hsqldb/types/IntervalMonthData.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2001-2011, The HSQL Development Group
+/* Copyright (c) 2001-2016, The HSQL Development Group
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -91,7 +91,7 @@ public class IntervalMonthData {
     }
 
     public int hashCode() {
-        return (int) units;
+        return units;
     }
 
     public int compareTo(IntervalMonthData b) {
